data: rvopcodesdecpder: log an error message prior to raising SystemExit
"rvopcodesdecoder" raises a SystemExit when a certain path is not found
while parsing for files in the opcodes directory.
It does so without printing any error messages, but just the error code "1".
This causes users to wonder what happened.
Fix this by logging an appropriate error message prior to raising an exit.

Signed-off-by: Karthik B K <karthik.bk@incoresemi.com>