data: rvopcodesdecoder: add support to change the path to opcodes_dir
Presently, "rvopcodesdecoder" does not allow the user to specify the path
to the opcodes directory. In cases where riscv-isac is imported as a python
module, this results in having a copy of riscv-opcodes, or the dev having to
figure out where isac is going to clone riscv-opcodes. Change this by
adding an argument to the setup and create_instr_dict methods which
has a default value 'None', that will lead to no change in behavior.

Signed-off-by: Karthik B K <karthik.bk@incoresemi.com>